OMAP: I2C: decouple spinlock init from usage in driver
If board file does not initialize hw-spinlock then we see error
with I2C driver read/write. This should not be the case. Even
without spinlock I2C driver should function as expected. This
patch addresses this issue and ensures that even if hw-spinlock
is not setup in board file, I2C driver is still functional

Reported-by: Dan Murphy <dmurphy@ti.com>
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>

Conflicts:
	drivers/i2c/busses/i2c-omap.c

Change-Id: I9b01d91f2108b0b948a6a78da16ed7ed3a6a53d2